Split up serialport package
