+++ pkg/serialport/fake.go
@@ -6,6 +6,7 @@ import (
 	"github.com/tarm/serial"
 )
 
+// MakeFakePort returns a mock serial port which behaves similarly to a real serial port and satisfies the Port interface
 func MakeFakePort(config *serial.Config) (*FakePort, error) {
 	return &FakePort{
 		config: config,
@@ -23,7 +24,7 @@ type FakePort struct {
 
 var responseMap = map[byte][]byte{
 	message.SingleParamMsg:      {message.OkStatus},
-	KeepAlive:                   {0x07},
+	message.KeepAlive:           {0x07},
 	message.CalibratedOffsetMsg: {0x21, 0x11, 0x30, 0x31, 0x46, 0x42, 0x46, 0x32},
 	message.CurrentOffsetMsg:    {0x21, 0x16, 0x30, 0x31, 0x46, 0x37, 0x46, 0x31},
 	message.DrillTypeMsg:        {0x21, 0x04, 0x35, 0x37, 0x33, 0x32, 0x34, 0x37, 0x34, 0x37, 0x33, 0x39, 0x33, 0x38, 0x33, 0x35, 0x33, 0x30, 0x30, 0x45},
@@ -39,8 +40,8 @@ func (mp *FakePort) Write(out []byte) (int, error) {
 	mp.Log.Println("Writing to mock port")
 	mp.writeLog = append(mp.writeLog, out...)
 	// If it's 0x07, we respond with the same
-	if len(out) == 1 && out[0] == byte(KeepAlive) {
-		mp.readBuffer = append(mp.readBuffer, KeepAlive)
+	if len(out) == 1 && out[0] == byte(message.KeepAlive) {
+		mp.readBuffer = append(mp.readBuffer, message.KeepAlive)
 		return len(out), nil
 	}
 	// we always read the second byte to see what to do
